requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
azure-ai-inference==1.0.0b4
azure-core==1.30.2
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boto3==1.38.26
botocore==1.38.26
browser-use==0.1.41
cachetools==5.5.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.1
courlan==1.3.2
cryptography==41.0.7
dataclasses-json==0.6.7
dateparser==1.2.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.2
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.0
djangorestframework==3.16.0
executing==2.0.1
faiss-cpu==1.11.0
fastjsonschema==2.19.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.3
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.32.2
hyperframe==6.1.0
ibm-cos-sdk==2.13.5
ibm-cos-sdk-core==2.13.5
ibm-cos-sdk-s3transfer==2.13.5
ibm_watsonx_ai==1.3.23
idna==3.6
iniconfig==2.1.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
joblib==1.5.1
json5==0.9.14
json_repair==0.46.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jusText==3.0.2
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-community==0.3.20
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ibm==0.3.10
langchain-mcp-adapters==0.0.9
langchain-mistralai==0.2.4
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.34
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.70
langsmith==0.3.43
lomond==0.3.3
lxml==5.4.0
lxml_html_clean==0.4.2
MainContentExtractor==0.0.4
Mako==1.3.10
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.4
marshmallow==3.26.1
matplotlib-inline==0.1.6
mcp==1.6.0
mcp_server_browser_use==0.1.8
mdurl==0.1.2
mem0ai==0.1.88
mistune==3.0.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.5
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
ollama==0.5.0
openai==1.82.1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pendulum==3.0.0
pillow==11.2.1
platformdirs==4.1.0
playwright==1.52.0
pluggy==1.6.0
portalocker==2.10.1
posthog==3.25.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.17.2
PyJWT==2.8.0
pyperclip==1.9.0
pytest==8.4.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.14.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2024.11.6
requests==2.32.3
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.17.1
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
screeninfo==0.8.1
Send2Trash==1.8.2
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
simple-salesforce==1.12.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.47.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.0
threadpoolctl==3.6.0
tiktoken==0.9.0
time-machine==2.13.0
tinycss2==1.2.1
tld==0.13.1
tokenizers==0.21.1
torch==2.7.0
tornado==6.4
tqdm==4.66.5
trafilatura==2.0.0
traitlets==5.14.1
transformers==4.52.3
typer==0.16.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.1.0
uv==0.7.8
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
widgetsnbextension==4.0.9
xxhash==3.5.0
yarl==1.20.0
zeep==4.2.1
zstandard==0.23.0
